Inline part of the TaskBuilder hierarchy; move TwoStageUniform to 'random' package.
